requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
